// types/supabase.ts
export interface Database {
  public: {
    Tables: {
      users: {
        Row: {
          id: string
          email: string
          role: 'Administrator' | 'Worker' | 'Customer'
          name: string | null
          created_at: string
        }
        Insert: {
          id?: string
          email: string
          role: 'Administrator' | 'Worker' | 'Customer'
          name?: string | null
          created_at?: string
        }
        Update: {
          id?: string
          email?: string
          role?: 'Administrator' | 'Worker' | 'Customer'
          name?: string | null
          created_at?: string
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}